Fix calculations without empty files and rmove to reports.
